Fixes async download.
