+++ lizard_connector/__init__.py
@@ -5,3 +5,4 @@ from lizard_connector import queries
 from lizard_connector import parsers
 
 from lizard_connector.connector import Client
+from lizard_connector import callbacks
